requirements.txt
python-telegram-bot[job-queue]
google-genai
python-dotenv